Move file transfer service to UnixSystems namespace
