+++ src/CommonServices/ServiceCollectionExtensions.cs
@@ -1,3 +1,4 @@
+using Kurmann.Videoschnitt.LocalFileSystem.UnixSystems;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Kurmann.Videoschnitt.CommonServices;
